Implement backup and restore (#446)
* Started stubbing a backup feature

* Add Tag Explorer page (#447)

* Implemented basic database backend of tags explorer

* Added navigation for new Tags page

* Added API for getting all tags

* Use ion-router-outlet instead of ion-nav

* Stubbed tags page

* Fix indentation

* A bit more complete

* Update page-tags.tsx

* Update page-tags.tsx

* Update page-tags.tsx

* Update page-tags.tsx

* Add sorting

* Fix sort button click

* Slight layout adjustments

* Add a page title

* Update app-root.tsx

* Change search filter icon

* Show less links on the toolbar

* Update app-root.tsx

* Improve side menu

* Update app-root.tsx

* Update app-root.tsx

* Update page-admin.tsx

* Update page-home.tsx

* Add sorting by tag

* Add unit tests

* Address sonar issues

* Update menu icon for tags

* onTagClicked does not need to be async

* Remove unecessary error case

* Bump version

* Fix uploading artifacts to release

* Show total number of recipes in the main menu (#448)

* Updated menu to show total recipe count

* Show cancel button when search is not default

* All states are included when no states filter is applied

* Include archived when selecting a tag so all matches show

* No need to call method twice

* Add tests for Find method

* Change iconn for reseting search filter

* Add unit tests for Find API route

* Address minor maintainability issue from Sonar

* Fix menu bugs

* Update Dockerfile to latest alpine 3

* Increase number of search results that can be seen per page (#450)

* Increase number of search results that can be seen per page

* Slight size adjustment

* Increase number of recipes in compact view also

* Fix compact layout

* Eliminate unecessary change in whitespace

* Allow selecting number of results per page (#451)

* Allow selecting number of results per page

* Use caret icon to match ion-select

* Undo icon swap

* Fix size mismatch when there is no image

* Fix unit test

* Interactivity animations and other layout tweaks (#453)

* Add hover zoom

* Tweak menu and toolbar

* Minor cleanup

* Bring back animations

* Fix unit tests

* Avoid 4 wide since we only show 6 results

* Move zoom to global styles

* Numerous layout tweaks related to cards and buttons

* Fix unit tests

* Adjust size of note title

* Date format tweaks

* Fix indentation

* Recipe viewer layout refresh (#454)

* WIP recipe viewer layout

* Add missing closing brace

* Update recipe-viewer.css

* Update recipe-viewer.css

* Update recipe-viewer.tsx

* Update recipe-viewer.css

* Update recipe-viewer.tsx

* Move servings

* Move time to header

* Update recipe-viewer.css

* Fix some failing test conditions

* Update tests

* Fix compile errors

* Update recipe-viewer.spec.tsx

* Update recipe-viewer.spec.tsx

* Adjust note title font size

* Bring back the helper text

* Adjust some card layouts

* Slightly increase image size

* Tweak home page layout

* Slight font size adjustments

* Slight font-size adjustment

* Usage ion-img for lazy loading; switch to thumbnail instead of avatar (#455)

* Usage ion-img for lazy loading; switch to thumbnail instead of avatar

* Load thumbnail initially

* Fix tests

* Attempt to recover some vertical space

* Slight tweak

* Make top image clickable

* Slightly taller main image in xl screens

* Remove buttons that are always hidden

* Fix space

* Remove extraneous class

* Implement more of backup database query

* Add placeholder to creating backup

* Partial implementatiion of recipe backup

* Refactor upload package to general file access

* Copy images into backup

* Backup users also

* Write backups toa zip file

* Use constant for base path

* Small cleanup

* Removed unused method

* Update for new env

* Fix unit test mocks

* Added some helper utilities

* Update backup to be all tables. Implement db import

* Add unit test for db Export

* Ensure consistent order of row values

* Add unit test for Import

* Create the backup on disk instead of in memory

* Remove references to old upload driver

* Another obsolete reference

* Still another one

* Pass context

* Fix bug in path and URL construction

* Support both SQLite and PostgreSQL for backup

* Use local temp dir

* Add some debug code

* Write to backup directly

* Use alpine again, to support execing into the container

* Implemented more of the backup functionality

* Remove unused sendDeactivatingCallback

* Fix tests

* Simplify CopyDirectoryToZip


* Implement initial version of deleting a backup


* Remove obsolete nosec

* Convert image-upload-browser to generic file upload browser


* Update UI approach

* Fix folder structure of uploads in backup


* Implement backup metadata


* Built out most of the rest of backup and restore


* Implement the database import part of restore


* Simplification

* Fix mock


* Fix bare file uploads


* Code cleanup


* Configure revive

* Small clarity cleanup


* Small cleanup


* Small consistency change

* Oops, missed a file

* Show backup size


* Fix buttons on smaller screens


* Handle DB special cases


* Handle triggers and constraints on import


* 750 permissions

* Eliminate need for Trap method


* Fix mocking

* Add tests for zip methods


* Added tests for API layer


* Fix code smell

* Test and fix unbuferedReaderAt


* Code smells


* More code smells


* Add version to UI


* Reorder restore for better error handling


* Allow alert to close before executing selected option


* Fix log message


* Consolidate logging infra for contexts


* Fix several tests that were susceptible to ordering issues


* Tests for onlyFilesFileSystem


* Fix some contexts in tests

* Convert prompts to skills


* Version bump

---------

+++ static/src/components/pages/admin-users/page-admin-users.tsx
@@ -2,13 +2,13 @@ import { alertController, modalController } from '@ionic/core';
 import { Component, Element, Host, h, State, Method } from '@stencil/core';
 import { User } from '../../../generated';
 import { usersApi } from '../../../helpers/api';
-import { enableBackForOverlay, isNull, showToast } from '../../../helpers/utils';
+import { ComponentWithActivatedCallback, enableBackForOverlay, isNull, showToast } from '../../../helpers/utils';
 
 @Component({
   tag: 'page-admin-users',
   styleUrl: 'page-admin-users.css',
 })
-export class PageAdminUsers {
+export class PageAdminUsers implements ComponentWithActivatedCallback {
   @State() users: User[] = [];
 
   @Element() el!: HTMLPageAdminUsersElement;
@@ -145,21 +145,19 @@ export class PageAdminUsers {
         header: 'Delete User?',
         message: `Are you sure you want to delete ${user.username}?`,
         buttons: [
-          'No',
-          {
-            text: 'Yes',
-            handler: async () => {
-              await this.deleteUser(user);
-              await this.loadUsers();
-              return true;
-            }
-          }
+          { text: 'No', role: 'cancel' },
+          { text: 'Yes', role: 'confirm' }
         ],
       });
 
       await confirmation.present();
 
-      await confirmation.onDidDismiss();
+      const { role } = await confirmation.onDidDismiss();
+
+      if (role === 'confirm') {
+        await this.deleteUser(user);
+        await this.loadUsers();
+      }
     });
   }
